+++ Classes/Form/Element/AdvancedRobots.php
@@ -5,19 +5,9 @@ declare(strict_types=1);
 namespace YoastSeoForTypo3\YoastSeo\Form\Element;
 
 use TYPO3\CMS\Backend\Form\AbstractNode;
-use TYPO3\CMS\Backend\Form\NodeFactory;
 
 class AdvancedRobots extends AbstractNode
 {
-    /**
-     * @param NodeFactory $nodeFactory
-     * @param array $data
-     */
-    public function __construct(NodeFactory $nodeFactory, array $data)
-    {
-        parent::__construct($nodeFactory, $data);
-    }
-
     public function render(): array
     {
         $resultArray = $this->initializeResultArray();